requirements.txt
# LangChain Core
langchain
langchain-core
langchain_experimental
langchain-chroma
chromadb
faiss-cpu
tiktoken
wikipedia

# Youtube transcript
youtube_transcript_api

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# For Structured output
pydantic

# For app
streamlit
